DELETED: useless code, changed name of loan service. impl for classes is the STANDARD.
