+++ src/main/java/net/unir/missi/desarrollowebfullstack/bookabook/operador/repository/LoanRepository.java
@@ -22,52 +22,44 @@ public class LoanRepository {
         this.loanJpaRepository = loanJpaRepository;
     }
 
+    public List<Loan> search(Long bookId, Long clientId, Date loanDate, Date returnDate, Date dueDate, Boolean isReturned, Integer renewalCount) {
+        SearchCriteria<Loan> spec = new SearchCriteria<>();
+
+        if(bookId != null) {
+            if (bookId != 0) {
+                spec.add(new SearchStatement("bookId", bookId, SearchOperation.EQUAL));
+            }
+        }
+        if(clientId != null) {
+            if (clientId != 0) {
+                spec.add(new SearchStatement("clientId", clientId, SearchOperation.EQUAL));
+            }
+        }
+        if(loanDate != null){
+            spec.add(new SearchStatement("loanDate", loanDate, SearchOperation.EQUAL));
+        }
+        if(returnDate != null){
+            spec.add(new SearchStatement("returnDate", returnDate, SearchOperation.EQUAL));
+        }
+        if(dueDate != null){
+            spec.add(new SearchStatement("dueDate", dueDate, SearchOperation.EQUAL));
+        }
+        if(isReturned != null){
+            spec.add(new SearchStatement("isReturned", isReturned, SearchOperation.EQUAL));
+        }
+        if(renewalCount != null){
+            spec.add(new SearchStatement("renewalCount", renewalCount, SearchOperation.EQUAL));
+        }
+
+        List<Loan> loanList = loanJpaRepository.findAll(spec);
+        return loanList;
+    }
+
     public List<Loan> findAll() { return loanJpaRepository.findAll(); }
     public Loan findById(Long id) { return loanJpaRepository.findById(id).orElse(null); }
     public List<Loan> findByClientId(Long clientId){ return loanJpaRepository.findByClientId(clientId); }
     public Loan save(Loan loan){ return loanJpaRepository.save(loan); }
     public void delete(Loan loan) { loanJpaRepository.delete(loan); }
 
-    public List<Loan> search(Long bookId, Long clientId, Date loanDate, Date returnDate, Date dueDate, Boolean isReturned, Integer renewalCount) {
-        SearchCriteria<Loan> spec = new SearchCriteria<>();
 
-        if(bookId != null) {
-            if (bookId != null && bookId != 0) {
-                spec.add(new SearchStatement("bookId", bookId, SearchOperation.EQUAL));
-            }
-        }
-        if(clientId != null) {
-            if (clientId != null && clientId != 0) {
-                spec.add(new SearchStatement("clientId", clientId, SearchOperation.EQUAL));
-            }
-        }
-        if(loanDate != null){
-            if (loanDate != null) {
-                spec.add(new SearchStatement("loanDate", loanDate, SearchOperation.EQUAL));
-            }
-        }
-        if(returnDate != null){
-            if (returnDate != null) {
-                spec.add(new SearchStatement("returnDate", returnDate, SearchOperation.EQUAL));
-            }
-        }
-        if(dueDate != null){
-            if (dueDate != null) {
-                spec.add(new SearchStatement("dueDate", dueDate, SearchOperation.EQUAL));
-            }
-        }
-        if(isReturned != null){
-            if (isReturned != null) {
-                spec.add(new SearchStatement("isReturned", isReturned, SearchOperation.EQUAL));
-            }
-        }
-        if(renewalCount != null){
-            if (renewalCount != null) {
-                spec.add(new SearchStatement("renewalCount", renewalCount, SearchOperation.EQUAL));
-            }
-        }
-
-        List<Loan> loanList = loanJpaRepository.findAll(spec);
-        return loanList;
-    }
 }
